Update Docker Hub API Call
Docker Hub API v1 is not available anymore. This updates the API call we
need to v2. Also includes minor formatting changes, and the addition of
a logo to the README, as well as a move of the UNLICENSE Markdown file
to standard UNLICENSE file so that tools can recognize the license.